refactor: split rooms into MVVM, fix Texture patterns and nested nav bar
