+++ Zyna/Application/SceneDelegate.swift
@@ -16,8 +16,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window = UIWindow(windowScene: windowScene)
         guard let window = self.window else { return }
 
+        appCoordinator.window = window
         appCoordinator.start()
-        window.rootViewController = appCoordinator.navigationController
         window.makeKeyAndVisible()
     }
 }
